requirements.txt
boto3>=1.34.0
dramatiq[redis]>=1.18.0
fastapi>=0.110.0
fpdf2>=2.7.9
Jinja2>=3.1
markdown>=3.6
pydantic>=2,<3
python-dotenv>=1.0.0
redis>=5.0
requests>=2.31.0
typer>=0.9.0
uvicorn>=0.30.0
weasyprint>=62.0
